+++ merchant/src/main/java/com/shopflix/merchant/data/ProductVariantData.java
@@ -7,7 +7,7 @@ public class ProductVariantData
 {
     private Long id;
     private String name;
-    private Integer quantity;
+    private Long quantity;
     private BigDecimal price;
     private String sku;
     private String value1;
@@ -34,12 +34,12 @@ public class ProductVariantData
         this.name = name;
     }
 
-    public Integer getQuantity()
+    public Long getQuantity()
     {
         return quantity;
     }
 
-    public void setQuantity(Integer quantity)
+    public void setQuantity(Long quantity)
     {
         this.quantity = quantity;
     }
